Authentication

+++ user.go
@@ -14,7 +14,7 @@ type UserService interface {
 }
 
 type User struct {
-	ID        uint64    `json:"id"`
+	ID        int       `json:"id"`
 	Username  string    `json:"username"`
 	Email     string    `json:"email"`
 	Password  string    `json:"-"`
